add 2 modules
