+++ index.html
@@ -248,7 +248,7 @@
     </div>
   </template>
 
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
 </body>
 
 </html>
